chore: upgrade to Java 21 docker image
Eclipse Temurin is the distribution we build with.
It is a natural choice, having not found an OpenJDK JRE.

The project has changed to using Java 17.
It would have been possible to use 17 but Java21 appears to work.
This makes the build ready for changing the source version.

TIS21-5226: Disconnect Doctors